refactor(server): extract task document persistence
